requirements.txt
fastmcp
pymongo
python-dotenv
requests
groq
openai

